New: moved new function into utils folder
